backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-ai
openai>=1.6.1
python-multipart==0.0.12
python-dotenv==1.0.1
weave
pytest==8.3.4
pytest-asyncio==0.24.0
sqlmodel==0.0.21
psycopg2-binary==2.9.9
pgvector==0.2.3
asyncpg==0.29.0
alembic==1.13.1 